third_party/blink/web_tests/platform/win/external/wpt/editing/run/strikethrough_2001-last-expected.txt

This is a testharness.js-based test.
Found 21 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandState("strikethrough") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo<s>[bar]</s>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo{<s>bar</s>}baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "foo{<s>bar</s>}baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo{<s>bar</s>}baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "foo{<s>bar</s>}baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "fo[o<span style=text-decoration:line-through>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<strike>o</strike><s>bar</s>baz" but got "fo<strike>ob</strike><span style=\\"text-decoration:line-through\\">ar</span>baz"
[FAIL] [["strikethrough",""]] "fo[o<span style=text-decoration:line-through>b]ar</span>baz" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["strikethrough",""]] "fo[o<span style=text-decoration:line-through>b]ar</span>baz" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>fo[o</strike><s>b]ar</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<strike>fo[o</strike><s>b]ar</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>fo[o</strike><s>b]ar</s>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<strike>fo[o</strike><s>b]ar</s>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandValue("strikethrough") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["strikethrough",""]] "<s>fo[o</s><del>b]ar</del>" queryCommandValue("strikethrough") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "abc<span style=\\"text-decoration:blink overline underline\\">[def]</span>ghi" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "abc<span style=\\"text-decoration-line:underline overline line-through blink\\">def</span>ghi" not in array ["abc<span style=\\"text-decoration:underline overline line-through\\">def</span>ghi", "abc<span style=\\"text-decoration-line:underline overline line-through\\">def</span>ghi"]
[FAIL] [["stylewithcss","true"],["strikethrough",""]] "abc<span style=\\"text-decoration:blink line-through underline overline\\">[def]</span>ghi" compare innerHTML
  assert_in_array: Unexpected innerHTML (after normalizing inline style) value "abc<span style=\\"text-decoration-line:underline overline blink\\">def</span>ghi" not in array ["abc<span style=\\"text-decoration:underline overline\\">def</span>ghi", "abc<span style=\\"text-decoration-line:underline overline\\">def</span>ghi"]
Harness: the test ran to completion.
